+++ src/main/java/com/alipay/global/api/request/ams/risk/tee/enums/ErrorCodeEnum.java
@@ -8,7 +8,7 @@
  * THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE; and 2. If applicable, the use of the Software is also subject to the terms and
  * conditions of any non-disclosure agreement signed by you and the relevant Ant Group entity.
  */
-package com.alipay.global.api.request.ams.risk.crypto.enums;
+package com.alipay.global.api.request.ams.risk.tee.enums;
 
 public enum ErrorCodeEnum {
